pyproject.toml
[tool.poetry]
name = "ec2-slackbot"
version = "0.1.1"
description = ""
authors = ["Robert Dargavel Smith <teticio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
slack-sdk = "^3.29.0"
boto3 = "^1.34.131"
pyyaml = "^6.0.1"
requests = "^2.32.3"
cryptography = "^42.0.8"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
coverage = "^7.5.4"
paramiko = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ec2-slackbot = "ec2_slackbot.app:main"
